refactor: App 컴포넌트에서 setProducts 제거 및 useProductStorage 훅을 통해 상품 관리 로직 통합
